+++ src/balancer/agent_controller_snapshot.rs
@@ -6,5 +6,5 @@ pub struct AgentControllerSnapshot {
     pub id: String,
     pub name: Option<String>,
     pub slots_processing: i32,
-    pub slots_total: usize,
+    pub slots_total: i32,
 }
