+++ src/balancer/management_service/http_route/api/ws_agent_socket/jsonrpc/notification_params/register_agent_params.rs
@@ -4,5 +4,5 @@ use serde::Serialize;
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
 pub struct RegisterAgentParams {
     pub name: Option<String>,
-    pub slots_total: usize,
+    pub slots_total: i32,
 }
